+++ inc/Acf/OptionsPage.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Airfleet\Framework\Acf;
+
+use Airfleet\Framework\Features\Feature;
+
+/**
+ * Create an ACF options page.
+ */
+class OptionsPage implements Feature {
+	protected array $options;
+
+	public function __construct( array $options ) {
+		$this->options = $options;
+	}
+
+	public function initialize(): void {
+		add_action(
+			'acf/init',
+			function () {
+				\acf_add_options_page( $this->options );
+			}
+		);
+	}
+}