+++ inc/Plugin/ActionLinks.php
@@ -32,7 +32,7 @@ class ActionLinks implements Feature {
 	/**
 	 * Constructor.
 	 *
-	 * @param array $options Enqueue properties.
+	 * @param array $options Plugin properties.
 	 *   $options = [
 	 *     'slug'  => (string) The slug
 	 *     'path'  => (string) Path to the base folder (plugin/theme)